refactor: 포트폴리오 API 컴포넌트와 같이 분리
