+++ src/components/Portfolio/PortfolioHoldings/IconCalendar.tsx
@@ -8,7 +8,7 @@ export const IconCalendar = () => {
       component={CalendarIcon}
       width={14}
       height={14}
-      viewBox="0 0 14 14"
+      inheritViewBox
       htmlColor={designSystem.color.neutral.gray600}
     />
   );
